refactor(chat): split panel presentation
